move cosigner folder
